requirements.txt
haystack-ai==2.10.3
datasets==3.3.2
sentence-transformers==3.4.1
google-ai-haystack==5.1.0
fastapi==0.115.4
uvicorn==0.31.0
beautifulsoup4==4.12.0  # Stable HTML parsing
umap-learn==0.5.4
plotly==5.22.0
scikit-learn==1.4.1.post1
numpy>=1.20.0
pandas>=1.3.0